reading temperature from micro
